fix: Lombok compiler warning
The compilation process was complaining about not properly specified defaults.

The `asyncapi` values are supposed to be `final`, but that change breaks the
`AsyncAPI` contract usage in different ways. The alternative is to indicate
the value is the default one for the builder. This solution removes the
warning while keeping the contract expectations.