fix: properly support multiple types
Some of the AsyncAPI fields support more than an object type. Previously this problem was
"solved" accepting any `Object` but this approach required castings or directly losing
the generic validations in any consumer code.

Here we propose to define an interface for each field type. For example `MessageBindingValue`
interface can be implemented both by the `Message` and the `Reference` models.

Note this PR is WIP to discuss the approach before making it more generic.